# backend/grpc-gateway/pkg/protocol/rest/openapi/url-shortener-service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "url-shortener-service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UrlShortenerService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/shortens/account-gen": {
      "post": {
        "operationId": "UrlShortenerService_generateShortCodeAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gonlinkGenerateShortCodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gonlinkGenerateShortCodeAccountRequest"
            }
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    },
    "/api/v1/shortens/gen": {
      "post": {
        "operationId": "UrlShortenerService_generateShortCode",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gonlinkGenerateShortCodeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gonlinkGenerateShortCodeRequest"
            }
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    },
    "/api/v1/shortens/get": {
      "post": {
        "operationId": "UrlShortenerService_getOriginalUrl",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/gonlinkGetOriginalUrlResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/gonlinkGetOriginalUrlRequest"
            }
          }
        ],
        "tags": [
          "UrlShortenerService"
        ]
      }
    }
  },
  "definitions": {
    "gonlinkGenerateShortCodeAccountRequest": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        },
        "trafficDate": {
          "type": "string"
        },
        "zoneId": {
          "type": "string"
        }
      }
    },
    "gonlinkGenerateShortCodeRequest": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        },
        "trafficDate": {
          "type": "string"
        },
        "zoneId": {
          "type": "string"
        }
      },
      "title": "=== GenerateShortCode ==="
    },
    "gonlinkGenerateShortCodeResponse": {
      "type": "object",
      "properties": {
        "shortCode": {
          "type": "string"
        },
        "base64Image": {
          "type": "string"
        },
        "isOwner": {
          "type": "boolean"
        }
      }
    },
    "gonlinkGetOriginalUrlRequest": {
      "type": "object",
      "properties": {
        "shortCode": {
          "type": "string"
        },
        "clientTime": {
          "type": "string"
        },
        "zoneId": {
          "type": "string"
        }
      },
      "title": "=== GetOriginalUrl ==="
    },
    "gonlinkGetOriginalUrlResponse": {
      "type": "object",
      "properties": {
        "originalUrl": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}